refactor(sample/11) change sample to reflect real world usage + typed responses
