+++ sample/11-swagger/src/cats/cats.service.ts
@@ -1,15 +1,23 @@
 import { Injectable } from '@nestjs/common';
-import { Cat } from './interfaces/cat.interface';
+import { Cat } from './classes/cat.class';
+import { CreateCatDto } from './dto/create-cat.dto';
 
 @Injectable()
 export class CatsService {
-  private readonly cats: Cat[] = [];
+  private readonly cats: CreateCatDto[] = [];
 
-  create(cat: Cat) {
+  create(cat: CreateCatDto): Cat {
     this.cats.push(cat);
+    return {
+      id: (this.cats.length - 1),
+      ...cat,
+    };
   }
 
   findOne(id: number): Cat {
-    return this.cats[id];
+    return {
+      id,
+      ...this.cats[id],
+    };
   }
 }
