+++ sample/11-swagger/src/cats/classes/cat.class.ts
@@ -1,0 +1,15 @@
+import { ApiModelProperty } from '@nestjs/swagger';
+
+export class Cat {
+  @ApiModelProperty()
+  id: number;
+
+  @ApiModelProperty()
+  name: string;
+
+  @ApiModelProperty()
+  age: number;
+
+  @ApiModelProperty()
+  breed: string;
+}
