--- sample/11-swagger/src/cats/interfaces/cat.interface.ts
@@ -1,5 +1,0 @@
-export interface Cat {
-  readonly name: string;
-  readonly age: number;
-  readonly breed: string;
-}
